Extract product and category tabs into individual components
